feat: 添加几何体优化和权限管理接口，更新相关模块以支持新功能
