+++ packages/specification/src/workflow/README.md
@@ -1,0 +1,244 @@
+# Workflow 模块
+
+## 概述
+
+Workflow 模块定义了 Maxellabs 3D Engine 的工作流程管理系统，支持从设计、开发到发布的完整生产流程管理。
+
+## 主要功能
+
+### 工作流程定义
+- **流程模板**: 预定义的标准工作流程
+- **自定义流程**: 灵活的自定义工作流程定义
+- **步骤管理**: 详细的步骤定义和依赖关系
+- **条件控制**: 基于条件的流程分支和循环
+
+### 任务管理
+- **任务调度**: 自动化任务调度和执行
+- **并行处理**: 支持多任务并行执行
+- **依赖管理**: 任务间的依赖关系管理
+- **状态追踪**: 实时的任务状态监控
+
+### 权限管理
+- **角色定义**: 灵活的角色权限系统
+- **访问控制**: 细粒度的权限控制
+- **审批流程**: 多级审批和权限验证
+- **操作日志**: 完整的操作记录和审计
+
+### 集成能力
+- **版本控制**: Git集成和版本管理
+- **CI/CD**: 持续集成和部署支持
+- **通知系统**: 多渠道通知和提醒
+- **第三方集成**: 支持主流开发工具集成
+
+## 核心类型
+
+### 工作流程层次
+```typescript
+// 工作流程结构
+WorkflowDefinition
+├── WorkflowMetadata      // 工作流元数据
+├── WorkflowStep[]        // 工作流步骤
+├── WorkflowVariable[]    // 工作流变量
+├── WorkflowTrigger[]     // 触发器
+└── WorkflowValidation    // 验证规则
+```
+
+### 任务处理层次
+```typescript
+// 任务处理结构
+WorkflowExecution
+├── ExecutionContext      // 执行上下文
+├── StepExecution[]       // 步骤执行
+├── TaskQueue            // 任务队列
+└── ExecutionResult      // 执行结果
+```
+
+## 使用示例
+
+### 创建3D内容工作流（使用命名空间）
+```typescript
+import { Workflow } from '@maxellabs/specification';
+
+const designWorkflow: Workflow.WorkflowDefinition = {
+  id: 'design-to-production',
+  name: '3D内容设计到生产流程',
+  description: '从设计到最终发布的完整工作流程',
+  version: '1.0.0',
+  steps: [
+    {
+      id: 'design-phase',
+      name: '设计阶段',
+      type: Workflow.WorkflowTaskType.DesignCreation,
+      description: '3D场景设计和建模',
+      inputs: ['design-requirements'],
+      outputs: ['3d-models', 'textures'],
+      assignee: 'designer-role'
+    },
+    {
+      id: 'review-phase',
+      name: '审核阶段',
+      type: Workflow.WorkflowTaskType.DesignReview,
+      description: '设计审核和反馈',
+      inputs: ['3d-models'],
+      outputs: ['approval-result'],
+      assignee: 'reviewer-role'
+    }
+  ],
+  triggers: [
+    {
+      type: Workflow.WorkflowTriggerType.Manual,
+      event: 'design-complete'
+    }
+  ]
+};
+```
+
+### 配置自动化任务（使用命名空间）
+```typescript
+import { Workflow } from '@maxellabs/specification';
+
+const optimizationTask: Workflow.AutomationTask = {
+  id: 'asset-optimization',
+  name: '资产优化任务',
+  type: Workflow.WorkflowTaskType.Build,
+  description: '自动优化3D模型和纹理',
+  configuration: {
+    inputPath: '/assets/raw/',
+    outputPath: '/assets/optimized/',
+    optimization: {
+      geometry: {
+        simplification: true,
+        compression: true
+      },
+      texture: {
+        compression: ['DXT', 'ETC'],
+        maxSize: 2048
+      }
+    }
+  },
+  executionMode: 'automatic',
+  schedule: {
+    type: 'cron',
+    expression: '0 2 * * *' // 每天凌晨2点执行
+  }
+};
+```
+
+### 设置权限管理（使用命名空间）
+```typescript
+import { Workflow } from '@maxellabs/specification';
+
+const designerRole: Workflow.RoleDefinition = {
+  id: 'designer-role',
+  name: '3D设计师',
+  description: '负责3D内容创作和设计',
+  permissions: [
+    {
+      resource: 'assets',
+      actions: [Workflow.WorkflowPermission.Create, Workflow.WorkflowPermission.Read, Workflow.WorkflowPermission.Write],
+      accessLevel: 'full'
+    },
+    {
+      resource: 'workflows',
+      actions: [Workflow.WorkflowPermission.Read, Workflow.WorkflowPermission.Execute],
+      accessLevel: 'limited'
+    }
+  ],
+  restrictions: {
+    maxFileSize: '100MB',
+    allowedFormats: ['fbx', 'obj', 'gltf']
+  }
+};
+```
+
+## 工作流程模板
+
+### 游戏开发流程
+```
+概念设计 → 3D建模 → 纹理制作 → 场景搭建 → 优化测试 → 发布部署
+```
+
+### VR/AR项目流程
+```
+需求分析 → 原型设计 → 3D内容制作 → 交互开发 → 性能优化 → 设备测试 → 发布
+```
+
+### 产品可视化流程
+```
+产品建模 → 材质调整 → 渲染设置 → 后期处理 → 质量检查 → 客户确认 → 交付
+```
+
+## 设计原则
+
+1. **灵活性**: 支持多种工作流程模式和自定义
+2. **可视化**: 提供直观的流程图和状态展示
+3. **自动化**: 最大化自动化处理和减少人工干预
+4. **协作性**: 支持团队协作和多角色参与
+
+## 与其他模块的关系
+
+- **被 Package 使用**: 工作流程包含在包格式中（通过 Package 命名空间）
+- **集成 Design**: 管理设计文档的生命周期（通过 Design 命名空间）
+- **协调 Rendering**: 控制渲染任务的执行（通过 Rendering 命名空间）
+- **依赖 core**: 使用基础的状态和事件类型
+
+## 流程监控
+
+### 实时状态
+- 当前执行步骤
+- 任务进度百分比
+- 资源使用情况
+- 错误和警告信息
+
+### 历史记录
+- 执行历史日志
+- 性能统计数据
+- 失败分析报告
+- 用户操作审计
+
+### 通知系统
+- 邮件通知
+- Slack/Teams集成
+- 移动端推送
+- 自定义Webhook
+
+## 扩展机制
+
+### 自定义步骤
+```typescript
+interface CustomStep extends Workflow.WorkflowStep {
+  customHandler: string;
+  configuration: Record<string, any>;
+}
+```
+
+### 插件系统
+```typescript
+interface WorkflowPlugin {
+  install(workflow: WorkflowEngine): void;
+  uninstall(workflow: WorkflowEngine): void;
+}
+```
+
+## 命名空间使用
+
+从 v0.0.6 开始，Workflow 模块通过命名空间导出，避免与其他模块的命名冲突：
+
+```typescript
+// 推荐的导入方式
+import { Workflow } from '@maxellabs/specification';
+
+// 使用时加上命名空间前缀
+const workflow: Workflow.WorkflowDefinition = { ... };
+const permission = Workflow.WorkflowPermission.Read;
+
+// 与其他模块的协作
+import { Package, Design, Rendering } from '@maxellabs/specification';
+```
+
+## 注意事项
+
+- 工作流程支持热更新和版本管理
+- 权限系统与企业AD/LDAP集成
+- 支持分布式执行和负载均衡
+- 提供完整的API和SDK支持
